added the basis of stack navigator
