+++ package.json
@@ -5,6 +5,7 @@
     "expo": "^23.0.0",
     "react": "16.0.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-23.0.0.tar.gz",
-    "react-native-elements": "^0.18.4"
+    "react-native-elements": "^0.18.4",
+    "react-navigation": "^1.0.0-beta.21"
   }
 }
